+++ R/module_filter_data.R
@@ -52,7 +52,7 @@ srv_filter_data <- function(id, datasets, active_datanames, data_rv, is_active) 
 
 #' @rdname module_filter_data
 .make_filtered_teal_data <- function(modules, data, datasets = NULL, datanames) {
-  data <- eval_code(data, sprintf("%1$s._raw_ <- %1$s", datanames))
+  data <- eval_code(data, sprintf(".%1$s_raw <- %1$s", datanames))
   filtered_code <- teal.slice::get_filter_expr(datasets = datasets, datanames = datanames)
   filtered_teal_data <- .append_evaluated_code(data, filtered_code)
   filtered_datasets <- sapply(datanames, function(x) datasets$get_data(x, filtered = TRUE), simplify = FALSE)