+++ tests/testthat/test-module_teal.R
@@ -1,6 +1,6 @@
 # comment: srv_teal is exported so the tests here are extensive and cover srv_data as well.
 #          testing of srv_data is not needed.
-module_output_table <<- function(output, id) {
+module_summary_table <<- function(output, id) {
   testthat::skip_if_not_installed("rvest")
   table_id <- sprintf("teal_modules-%s-data_summary-table", id)
   html <- output[[table_id]]$html
@@ -405,9 +405,9 @@ testthat::describe("srv_teal teal_modules", {
         testthat::expect_null(modules_output$module_1())
 
         session$setInputs(`data-teal_data_module-data-dataset` = "iris", `teal_modules-active_tab` = "module_1")
-        testthat::expect_identical(
-          ls(teal.code::get_env(modules_output$module_1()())),
-          c("iris", "iris._raw_")
+        testthat::expect_setequal(
+          ls(teal.code::get_env(modules_output$module_1()()), all.names = TRUE),
+          c("iris", ".iris_raw")
         )
 
         # comment: can't trigger reactivity in testServer - the change in a reactive input data
@@ -415,9 +415,9 @@ testthat::describe("srv_teal teal_modules", {
         #          is sent to another teal_module
         session$setInputs(`data-teal_data_module-data-dataset` = "mtcars", `teal_modules-active_tab` = "module_2")
         session$flushReact()
-        testthat::expect_identical(
-          ls(teal.code::get_env(modules_output$module_2()())),
-          c("mtcars", "mtcars._raw_")
+        testthat::expect_setequal(
+          ls(teal.code::get_env(modules_output$module_2()()), all.names = TRUE),
+          c("mtcars", ".mtcars_raw")
         )
       }
     )
@@ -1228,11 +1228,11 @@ testthat::describe("srv_teal filters", {
           session$flushReact()
           # iris is not active
           testthat::expect_identical(modules_output$module_1()()[["iris"]], iris)
-          testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+          testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
           # mtcars has been modified
           expected_mtcars <- subset(mtcars, cyl == 4)
           testthat::expect_identical(modules_output$module_1()()[["mtcars"]], expected_mtcars)
-          testthat::expect_identical(modules_output$module_1()()[["mtcars._raw_"]], mtcars)
+          testthat::expect_identical(modules_output$module_1()()[[".mtcars_raw"]], mtcars)
 
           expected_code <- paste0(
             c(
@@ -1240,8 +1240,8 @@ testthat::describe("srv_teal filters", {
               "mtcars <- mtcars",
               sprintf('stopifnot(rlang::hash(iris) == "%s")', rlang::hash(iris)),
               sprintf('stopifnot(rlang::hash(mtcars) == "%s")', rlang::hash(mtcars)),
-              "iris._raw_ <- iris",
-              "mtcars._raw_ <- mtcars",
+              ".iris_raw <- iris",
+              ".mtcars_raw <- mtcars",
               "mtcars <- dplyr::filter(mtcars, cyl == 4)"
             ),
             collapse = "\n"
@@ -1380,9 +1380,9 @@ testthat::describe("srv_teal teal_module(s) transformer", {
       expr = {
         session$setInputs(`teal_modules-active_tab` = "module_1")
         testthat::expect_identical(modules_output$module_1()()[["iris"]], head(iris))
-        testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+        testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
         testthat::expect_identical(modules_output$module_1()()[["mtcars"]], head(mtcars))
-        testthat::expect_identical(modules_output$module_1()()[["mtcars._raw_"]], mtcars)
+        testthat::expect_identical(modules_output$module_1()()[[".mtcars_raw"]], mtcars)
       }
     )
   })
@@ -1415,17 +1415,17 @@ testthat::describe("srv_teal teal_module(s) transformer", {
         rownames(expected_iris) <- NULL
         expected_iris <- head(expected_iris)
         testthat::expect_identical(modules_output$module_1()()[["iris"]], expected_iris)
-        testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+        testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
         testthat::expect_identical(modules_output$module_1()()[["mtcars"]], head(subset(mtcars, cyl == 6)))
-        testthat::expect_identical(modules_output$module_1()()[["mtcars._raw_"]], mtcars)
+        testthat::expect_identical(modules_output$module_1()()[[".mtcars_raw"]], mtcars)
 
         expected_code <- paste(collapse = "\n", c(
           "iris <- iris",
           "mtcars <- mtcars",
           sprintf('stopifnot(rlang::hash(iris) == "%s")', rlang::hash(iris)),
           sprintf('stopifnot(rlang::hash(mtcars) == "%s")', rlang::hash(mtcars)),
-          "iris._raw_ <- iris",
-          "mtcars._raw_ <- mtcars",
+          ".iris_raw <- iris",
+          ".mtcars_raw <- mtcars",
           'iris <- dplyr::filter(iris, Species == "versicolor")',
           "mtcars <- dplyr::filter(mtcars, cyl == 6)",
           "iris <- head(iris, n = 6)",
@@ -1464,17 +1464,17 @@ testthat::describe("srv_teal teal_module(s) transformer", {
         session$flushReact()
 
         testthat::expect_identical(modules_output$module_1()()[["iris"]], head(iris))
-        testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+        testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
         testthat::expect_identical(modules_output$module_1()()[["mtcars"]], head(subset(mtcars, cyl == 4)))
-        testthat::expect_identical(modules_output$module_1()()[["mtcars._raw_"]], mtcars)
+        testthat::expect_identical(modules_output$module_1()()[[".mtcars_raw"]], mtcars)
 
         expected_code <- paste(collapse = "\n", c(
           "iris <- iris",
           "mtcars <- mtcars",
           sprintf('stopifnot(rlang::hash(iris) == "%s")', rlang::hash(iris)),
           sprintf('stopifnot(rlang::hash(mtcars) == "%s")', rlang::hash(mtcars)),
-          "iris._raw_ <- iris",
-          "mtcars._raw_ <- mtcars",
+          ".iris_raw <- iris",
+          ".mtcars_raw <- mtcars",
           "mtcars <- dplyr::filter(mtcars, cyl == 4)",
           "iris <- head(iris, n = 6)",
           "mtcars <- head(mtcars, n = 6)"
@@ -1582,7 +1582,7 @@ testthat::describe("srv_teal teal_module(s) transformer", {
       expr = {
         session$setInputs(`teal_modules-active_tab` = "module_1")
         testthat::expect_identical(modules_output$module_1()()[["iris"]], iris)
-        testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+        testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
       }
     )
   })
@@ -1611,7 +1611,7 @@ testthat::describe("srv_teal teal_module(s) transformer", {
       expr = {
         session$setInputs(`teal_modules-active_tab` = "module_1")
         testthat::expect_identical(modules_output$module_1()()[["iris"]], iris)
-        testthat::expect_identical(modules_output$module_1()()[["iris._raw_"]], iris)
+        testthat::expect_identical(modules_output$module_1()()[[".iris_raw"]], iris)
       }
     )
   })
@@ -1696,7 +1696,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("iris", "mtcars"),
             Obs = c("150/150", "32/32"),
@@ -1727,7 +1727,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("a", "b"),
             Obs = c("3/3", "6/6"),
@@ -1760,7 +1760,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("a", "b"),
             Obs = c("3/3", "6/6"),
@@ -1794,7 +1794,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("a", "b"),
             Obs = c("3/3", "6/6"),
@@ -1829,7 +1829,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("a", "b"),
             Obs = c("1/3", "2/6"),
@@ -1866,7 +1866,7 @@ testthat::describe("srv_teal summary table", {
         )
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("a", "b"),
             Obs = c("1/3", "2/6"),
@@ -1906,7 +1906,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("iris", "new_dataset"),
             Obs = c("150/150", "3"),
@@ -1936,7 +1936,7 @@ testthat::describe("srv_teal summary table", {
           session$setInputs("teal_modules-active_tab" = "module_1")
           session$flushReact()
           testthat::expect_identical(
-            module_output_table(output, "module_1"),
+            module_summary_table(output, "module_1"),
             data.frame(
               "Data Name" = c("iris"),
               Obs = c("6/150"),
@@ -1961,7 +1961,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("iris"),
             Obs = c("150/150"),
@@ -1978,7 +1978,6 @@ testthat::describe("srv_teal summary table", {
     teal.data::join_keys(data) <- teal.data::join_keys(
       teal.data::join_key("parent", "child", keys = c("am"))
     )
-
     shiny::testServer(
       app = srv_teal,
       args = list(
@@ -1990,7 +1989,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1")[["Data Name"]],
+          module_summary_table(output, "module_1")[["Data Name"]],
           c("parent", "child")
         )
       }
@@ -2010,7 +2009,7 @@ testthat::describe("srv_teal summary table", {
         session$setInputs("teal_modules-active_tab" = "module_1")
         session$flushReact()
         testthat::expect_identical(
-          module_output_table(output, "module_1"),
+          module_summary_table(output, "module_1"),
           data.frame(
             "Data Name" = c("iris"),
             Obs = c("150/150"),